Disable Windows launcher process
The launcher process on Windows exists to display a pre-XUL skeleton UI,
which shows the position of where browser UI elements would go before XUL
starts up. It's a nice feature by modern UI design standards, but we feel
it doesn't work out nicely with Nara's planned user interface, and have
therefore disabled it.

A bit of backstory: Mozilla's official handling of this is really bad. They
only use the launcher process on Windows, but of course they ship multiple
applications, so they needed to add a specific workaround for Thunderbird
(and Thunderbird-based applications like SeaMonkey).

I decided to clean up the official implementation for Nara.