Revise README
Since Nara is hosted exclusively on GitHub, I decided to move the README to
a markdown file. The README additionally is revised to contain information
specifically pertinent to Nara, and be a little bit more accessible for
newcomers - although I don't necessarily suspect anyone new is going to be
looking into joining Nara development at this time.

Signed-off-by: Isabella <165211464+kawapure@users.noreply.github.com>