Fix -moz-image-region for <xul:image>
There was a small mistake regarding arguments in nsLayoutUtils.cpp. The
call should have been adjusted to use the reverted local variable `image`,
rather than the argument variable `aImage` when calling DrawImageInternal.

I also cleaned up the implementation specific to nsImageFrame.cpp (which is
what handles XUL image frames since Gecko 112). The image region should now
be applied roughly around the same time as the computed style is updated.

No reimplementation of nsImageBoxFrame was required for this go. I hope it
doesn't creep back up in the future, but if this was the only feature that
we needed to add it for, then it should be alright.

Signed-off-by: Isabella <165211464+kawapure@users.noreply.github.com>